\name{aardvark-package}
\alias{aardvark-package}
\alias{aardvark}
\docType{package}
\title{
\packageTitle{aardvark}
}
\description{
\packageDescription{aardvark}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{aardvark}
\packageIndices{aardvark}

The aardvark project is an R package management system with two primary goals:
completely automated package submission and package version management.
}
\author{
\packageAuthor{Jeremy Stephens}

Maintainer: \packageMaintainer{Jeremy Stephens}
}
\keyword{ package }
